+++ lib/index.ts
@@ -1,4 +1,9 @@
-export { Server, type ServerOptions, type DegradationEvent } from "./server";
+export {
+  Server,
+  type ServerOptions,
+  type DegradationEvent,
+  type Middleware,
+} from "./server";
 export { Socket, type CloseReason } from "./socket";
 export { type RawData, type Packet, type PacketType } from "./parser";
 export { type BunWebSocket, type WebSocketData } from "./transports/websocket";